Modularized to base.py
now it's easier to build the semaphore class
